Switched cost values from type int to long

+++ Jantimizer/QueryOptimiser/Cost/CostCalculators/ICostCalculator.cs
@@ -16,6 +16,6 @@ namespace QueryOptimiser.Cost.CostCalculators
     {
         public IHistogramManager<HistogramType, DbConnectorType> HistogramManager { get; set; }
 
-        public int CalculateCost(INode node);
+        public long CalculateCost(INode node);
     }
 }
